Isolated Bio.PDB DSSP tests
