+++ Tests/test_PDB.py
@@ -51,8 +51,6 @@ from Bio.PDB import HSExposureCA, HSExposureCB, ExposureCN
 from Bio.PDB.PDBExceptions import PDBConstructionException, PDBConstructionWarning
 from Bio.PDB import rotmat, Vector
 from Bio.PDB import Residue, Atom
-from Bio.PDB import make_dssp_dict
-from Bio.PDB import DSSP
 from Bio.PDB.NACCESS import process_asa_data, process_rsa_data
 from Bio.PDB.ResidueDepth import _get_atom_radius
 
@@ -1699,134 +1697,6 @@ class CopyTests(unittest.TestCase):
 def eprint(*args, **kwargs):
     """Print to stderr."""
     print(*args, file=sys.stderr, **kwargs)
-
-
-def will_it_float(s):
-    """Convert the input into a float if it is a number.
-
-    If the input is a string, the output does not change.
-    """
-    try:
-        return float(s)
-    except ValueError:
-        return s
-
-
-class DsspTests(unittest.TestCase):
-    """Tests for DSSP parsing etc which don't need the binary tool.
-
-    See also test_DSSP_tool.py for run time testing with the tool.
-    """
-
-    def test_DSSP_file(self):
-        """Test parsing of pregenerated DSSP."""
-        dssp, keys = make_dssp_dict("PDB/2BEG.dssp")
-        self.assertEqual(len(dssp), 130)
-
-    def test_DSSP_noheader_file(self):
-        """Test parsing of pregenerated DSSP missing header information."""
-        # New DSSP prints a line containing only whitespace and "."
-        dssp, keys = make_dssp_dict("PDB/2BEG_noheader.dssp")
-        self.assertEqual(len(dssp), 130)
-
-    def test_DSSP_hbonds(self):
-        """Test parsing of DSSP hydrogen bond information."""
-        dssp, keys = make_dssp_dict("PDB/2BEG.dssp")
-
-        dssp_indices = {v[5] for v in dssp.values()}
-        hb_indices = set()
-
-        # The integers preceding each hydrogen bond energy (kcal/mol) in the
-        # "N-H-->O    O-->H-N    N-H-->O    O-->H-N" dssp output columns are
-        # relative dssp indices. Therefore, "hb_indices" contains the absolute
-        # dssp indices of residues participating in (provisional) h-bonds. Note
-        # that actual h-bonds are typically determined by an energetic
-        # threshold.
-        for val in dssp.values():
-            hb_indices |= {val[5] + x for x in (val[6], val[8], val[10], val[12])}
-
-        # Check if all h-bond partner indices were successfully parsed.
-        self.assertEqual((dssp_indices & hb_indices), hb_indices)
-
-    def test_DSSP_in_model_obj(self):
-        """All elements correctly added to xtra attribute of input model object."""
-        p = PDBParser()
-        s = p.get_structure("example", "PDB/2BEG.pdb")
-        m = s[0]
-        # Read the DSSP data into the pdb object:
-        _ = DSSP(m, "PDB/2BEG.dssp", "dssp", "Sander", "DSSP")
-        # Now compare the xtra attribute of the pdb object
-        # residue by residue with the pre-computed values:
-        i = 0
-        with open("PDB/dssp_xtra_Sander.txt") as fh_ref:
-            ref_lines = fh_ref.readlines()
-            for chain in m:
-                for res in chain:
-                    # Split the pre-computed values into a list:
-                    xtra_list_ref = ref_lines[i].rstrip().split("\t")
-                    # Then convert each element to float where possible:
-                    xtra_list_ref = list(map(will_it_float, xtra_list_ref))
-                    # The xtra attribute is a dict.
-                    # To compare with the pre-computed values first sort according to keys:
-                    xtra_itemts = sorted(
-                        res.xtra.items(), key=lambda s: s[0]
-                    )  # noqa: E731
-                    # Then extract the list of xtra values for the residue
-                    # and convert to floats where possible:
-                    xtra_list = [t[1] for t in xtra_itemts]
-                    xtra_list = list(map(will_it_float, xtra_list))
-                    # The reason for converting to float is, that casting a float to a string in python2.6
-                    # will include fewer decimals than python3 and an assertion error will be thrown.
-                    self.assertEqual(xtra_list, xtra_list_ref)
-                    i += 1
-
-    def test_DSSP_RSA(self):
-        """Tests the usage of different ASA tables."""
-        # Tests include Sander/default, Wilke and Miller
-        p = PDBParser()
-        # Sander/default:
-        s = p.get_structure("example", "PDB/2BEG.pdb")
-        m = s[0]
-        # Read the DSSP data into the pdb object:
-        _ = DSSP(m, "PDB/2BEG.dssp", "dssp", "Sander", "DSSP")
-        # Then compare the RASA values for each residue with the pre-computed values:
-        i = 0
-        with open("PDB/Sander_RASA.txt") as fh_ref:
-            ref_lines = fh_ref.readlines()
-            for chain in m:
-                for res in chain:
-                    rasa_ref = float(ref_lines[i].rstrip())
-                    rasa = float(res.xtra["EXP_DSSP_RASA"])
-                    self.assertAlmostEqual(rasa, rasa_ref)
-                    i += 1
-
-        # Wilke (procedure similar as for the Sander values above):
-        s = p.get_structure("example", "PDB/2BEG.pdb")
-        m = s[0]
-        _ = DSSP(m, "PDB/2BEG.dssp", "dssp", "Wilke", "DSSP")
-        i = 0
-        with open("PDB/Wilke_RASA.txt") as fh_ref:
-            ref_lines = fh_ref.readlines()
-            for chain in m:
-                for res in chain:
-                    rasa_ref = float(ref_lines[i].rstrip())
-                    rasa = float(res.xtra["EXP_DSSP_RASA"])
-                    self.assertAlmostEqual(rasa, rasa_ref)
-                    i += 1
-
-        # Miller (procedure similar as for the Sander values above):
-        s = p.get_structure("example", "PDB/2BEG.pdb")
-        m = s[0]
-        _ = DSSP(m, "PDB/2BEG.dssp", "dssp", "Miller", "DSSP")
-        i = 0
-        with open("PDB/Miller_RASA.txt") as fh_ref:
-            ref_lines = fh_ref.readlines()
-            for chain in m:
-                for res in chain:
-                    rasa_ref = float(ref_lines[i].rstrip())
-                    rasa = float(res.xtra["EXP_DSSP_RASA"])
-                    self.assertAlmostEqual(rasa, rasa_ref)
-                    i += 1
 
 
 class ResidueDepthTests(unittest.TestCase):
